move all NMF related things out of BaseData and into SplicingData only
