+++ flotilla/data_model/base.py
@@ -11,16 +11,14 @@ import seaborn as sns
 from sklearn.preprocessing import StandardScaler
 from scipy.cluster.vq import whiten
 
-from ..compute.decomposition import DataFramePCA, DataFrameNMF
+from ..compute.decomposition import DataFramePCA
 from ..compute.infotheory import binify, cross_phenotype_jsd, jsd_df_to_2d
 from ..compute.predict import PredictorConfigManager, \
     PredictorDataSetManager, CLASSIFIER
 from ..visualize.decomposition import DecompositionViz
-from ..visualize.generic import violinplot, nmf_space_transitions, \
-    simple_twoway_scatter
+from ..visualize.generic import violinplot, simple_twoway_scatter
 from ..visualize.network import NetworkerViz
 from ..visualize.predict import ClassifierViz
-from ..util import memoize, cached_property
 from ..compute.outlier import OutlierDetection
 
 MINIMUM_FEATURE_SUBSET = 20
@@ -1032,21 +1030,6 @@ class BaseData(object):
     def _thresh_float(df, f):
         return f * df.shape[0]
 
-    @cached_property()
-    def nmf(self):
-        data = self._subset(self.data)
-        return DataFrameNMF(self.binify(data).T, n_components=2)
-
-    # @memoize
-    def binned_nmf_reduced(self, sample_ids=None, feature_ids=None,
-                           data=None):
-        if data is None:
-            data = self._subset(self.data, sample_ids, feature_ids,
-                                require_min_samples=False)
-        binned = self.binify(data)
-        reduced = self.nmf.transform(binned.T)
-        return reduced
-
     def plot_feature(self, feature_id, sample_ids=None,
                      phenotype_groupby=None,
                      phenotype_order=None, color=None,
@@ -1108,127 +1091,6 @@ class BaseData(object):
                     continue
             sns.despine()
         fig.tight_layout()
-
-    def nmf_space_positions(self, groupby, n=20):
-        """Calculate NMF-space position of splicing events in phenotype groups
-
-        Parameters
-        ----------
-        groupby : mappable
-            A sample id to phenotype mapping
-        n : int or float
-            If int, then this is the absolute number of cells that are minimum
-            required to calculate modalities. If a float, then require this
-            fraction of samples to calculate modalities, e.g. if 0.6, then at
-            least 60% of samples must have an event detected for modality
-            detection
-
-        Returns
-        -------
-        df : pandas.DataFrame
-            A (n_events, n_groups) dataframe of NMF positions
-        """
-        grouped = self.singles.groupby(groupby)
-        if isinstance(n, int):
-            thresh = self._thresh_int
-        elif isinstance(n, float):
-            thresh = self._thresh_float
-
-        at_least_n_per_group_per_event = pd.concat(
-            [df.dropna(thresh=thresh(df, n), axis=1) for name, df in grouped])
-        at_least_n_per_group_per_event = at_least_n_per_group_per_event.dropna(
-            how='all', axis=1)
-        # at_least_n_per_group_per_event = grouped.transform(
-        #     lambda x: x if x.count() >= n else pd.Series(np.nan,
-        #                                                  index=x.index))
-        df = at_least_n_per_group_per_event.groupby(groupby).apply(
-            lambda x: self.binned_nmf_reduced(data=x)
-            if x.count().sum() > 0 else pd.DataFrame())
-        df = df.swaplevel(0, 1)
-        df = df.sort_index()
-        return df
-
-    def plot_nmf_space_transitions(self, feature_id, groupby,
-                                   phenotype_to_color,
-                                   phenotype_to_marker, order, ax=None,
-                                   xlabel=None, ylabel=None, n=20):
-        nmf_space_positions = self.nmf_space_positions(groupby, n=n)
-
-        nmf_space_transitions(nmf_space_positions, feature_id,
-                              phenotype_to_color,
-                              phenotype_to_marker, order,
-                              ax, xlabel, ylabel)
-
-    @staticmethod
-    def transition_distances(positions, transitions):
-        """Get NMF distance of features between phenotype transitions
-
-        Parameters
-        ----------
-        positions : pandas.DataFrame
-            A ((n_features, phenotypes), 2) MultiIndex dataframe of the NMF
-            positions of splicing events for different phenotypes
-        transitions : list of 2-string tuples
-            List of (phenotype1, phenotype2) transitions
-
-        Returns
-        -------
-        transitions : pandas.DataFrame
-            A (n_features, n_transitions) DataFrame of the NMF distances
-            of features between different phenotypes
-        """
-        positions_phenotype = positions.copy()
-        positions_phenotype.index = positions_phenotype.index.droplevel(0)
-        distances = pd.Series(index=transitions)
-        for transition in transitions:
-            try:
-                phenotype1, phenotype2 = transition
-                norm = np.linalg.norm(positions_phenotype.ix[phenotype2] -
-                                      positions_phenotype.ix[phenotype1])
-                # print phenotype1, phenotype2, norm
-                distances[transition] = norm
-            except KeyError:
-                pass
-        return distances
-
-    def nmf_space_transitions(self, groupby, phenotype_transitions, n=20):
-        """Get distance in NMF space of different splicing events
-
-        Parameters
-        ----------
-        groupby : mappable
-            A sample id to phenotype mapping
-        phenotype_transitions : list of str pairs
-            Which phenotype follows from one to the next, for calculating
-            distances between
-        n : int or float
-            If int, then this is the absolute number of cells that are minimum
-            required to calculate modalities. If a float, then require this
-            fraction of samples to calculate modalities, e.g. if 0.6, then at
-            least 60% of samples must have an event detected for modality
-            detection
-
-        Returns
-        -------
-        nmf_space_transitions : pandas.DataFrame
-            A (n_events, n_phenotype_transitions) sized DataFrame of the
-            distances of these events in NMF space
-        """
-        nmf_space_positions = self.nmf_space_positions(groupby, n=n)
-
-        # Take only splicing events that have at least two phenotypes
-        nmf_space_positions = nmf_space_positions.groupby(
-            level=0, axis=0).filter(lambda x: len(x) > 1)
-
-        nmf_space_transitions = nmf_space_positions.groupby(
-            level=0, axis=0, as_index=True, group_keys=False).apply(
-            self.transition_distances, transitions=phenotype_transitions)
-
-        # Remove any events that didn't have phenotype pairs from
-        # the transitions
-        nmf_space_transitions = nmf_space_transitions.dropna(how='all',
-                                                             axis=0)
-        return nmf_space_transitions
 
     def plot_two_samples(self, sample1, sample2, fillna=None,
                          **kwargs):
