+++ flotilla/data_model/expression.py
@@ -59,17 +59,6 @@ class ExpressionData(BaseData):
             standardize=standardize, metric=metric,
             linkage_method=linkage_method)
 
-    @memoize
-    def binify(self, data):
-        data = self._subset(data, require_min_samples=False)
-        data = (data - data.min()) / (data.max() - data.min())
-        # vmax = data.abs().max().max()
-        # vmin = -vmax
-        # bins = np.linspace(vmin, vmax, 10)
-        bins = np.arange(0, 1.1, .1)
-        # print 'bins:', bins
-        return super(ExpressionData, self).binify(data, bins)
-
 
 class SpikeInData(ExpressionData):
     """Class for Spikein data and associated functions
